Support disabling the parsing code
There are use cases for this crate in which the parsing functionality is not
needed. However, the parsing code still incurs a dependency cost in this case.
This commit adds a new cargo feature "parse" (on by default), which if disabled
means the crate can only be used for dumping but then has only a single
dependency.